Check that an argument is an atomic vector

An atomic vector is defined slightly different from specifications in is.atomic and is.vector: An atomic vector is either logical, integer, numeric, complex, character or raw and can have any attributes except a dimension attribute (like matrices). I.e., a factor is an atomic vector, but a matrix or NULL are not. In short, this is mostly equivalent to is.atomic(x) && !is.null(x) && is.null(dim(x)).

checkAtomicVector(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

check_atomic_vector(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

assertAtomicVector(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  .var.name = vname(x), add = NULL)

assert_atomic_vector(x, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, .var.name = vname(x), add = NULL)

testAtomicVector(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

test_atomic_vector(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

expect_atomic_vector(x, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, info = NULL, label = vname(x))

Arguments

x
[any] Object to check.

any.missing
[logical(1)] Are vectors with missing values allowed? Default is TRUE.

all.missing
[logical(1)] Are vectors with only missing values allowed? Default is TRUE.

len
[integer(1)] Exact expected length of x.

min.len
[integer(1)] Minimal length of x.

max.len
[integer(1)] Maximal length of x.

unique
[logical(1)] Must all values be unique? Default is FALSE.

names
[character(1)] Check for names. See checkNamed for possible values. Default is “any” which performs no check at all.

.var.name
[character(1)] Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add
[AssertCollection] Collection to store assertion messages. See AssertCollection.

info
[character(1)] Extra information to be included in the message for the testthat reporter. See expect_that.

label
[character(1)] Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertAtomicVector/assert_atomic_vector return x invisibly, whereas checkAtomicVector/check_atomic_vector and testAtomicVector/test_atomic_vector return TRUE. If the check is not successful, assertAtomicVector/assert_atomic_vector throws an error message, testAtomicVector/test_atomic_vector returns FALSE, and checkAtomicVector returns a string with the error message. The function expect_atomic_vector always returns an expectation.

See also

Other basetypes: checkArray, checkAtomic, checkCharacter, checkComplex, checkDataFrame, checkDate, checkEnvironment, checkFactor, checkFunction, checkIntegerish, checkInteger, checkList, checkLogical, checkMatrix, checkNull, checkNumeric, checkPOSIXct, checkVector

Other atomicvector: checkAtomic, checkVector

Examples

testAtomicVector(letters, min.len = 1L, any.missing = FALSE)

#> [1] TRUE