Check if an argument is a vector of type character

Check if an argument is a vector of type character

checkCharacter(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

check_character(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

assertCharacter(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE, .var.name = vname(x), add = NULL)

assert_character(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE, .var.name = vname(x), add = NULL)

testCharacter(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

test_character(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

expect_character(x, min.chars = NULL, pattern = NULL, fixed = NULL,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE, info = NULL, label = vname(x))

Arguments

x

[any] Object to check.

min.chars

[integer(1)] Minimum number of characters in each element of x.

pattern

[character(1L)] Regular expression as used in grepl. All elements of x must comply to this pattern.

fixed

[character(1)] Substring to detect in x. Will be used as pattern in grepl with option fixed set to TRUE. All elements of x must contain this substring.

ignore.case

[logical(1)] See grepl. Default is FALSE.

any.missing

[logical(1)] Are vectors with missing values allowed? Default is TRUE.

all.missing

[logical(1)] Are vectors with only missing values allowed? Default is TRUE.

len

[integer(1)] Exact expected length of x.

min.len

[integer(1)] Minimal length of x.

max.len

[integer(1)] Maximal length of x.

unique

[logical(1)] Must all values be unique? Default is FALSE.

names

[character(1)] Check for names. See checkNamed for possible values. Default is “any” which performs no check at all. Note that you can use checkSubset to check for a specific set of names.

null.ok

[logical(1)] If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name

[character(1)] Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add

[AssertCollection] Collection to store assertion messages. See AssertCollection.

info

[character(1)] Extra information to be included in the message for the testthat reporter. See expect_that.

label

[character(1)] Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertCharacter/assert_character return x invisibly, whereas checkCharacter/check_character and testCharacter/test_character return TRUE. If the check is not successful, assertCharacter/assert_character throws an error message, testCharacter/test_character returns FALSE, and checkCharacter returns a string with the error message. The function expect_character always returns an expectation.

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ NA_character_ and NaN.

See also

Other basetypes: checkArray, checkAtomic, checkComplex, checkDataFrame, checkDate, checkEnvironment, checkFactor, checkFunction, checkIntegerish, checkInteger, checkList, checkLogical, checkMatrix, checkNull, checkNumeric, checkVector

Examples

testCharacter(letters, min.len = 1, any.missing = FALSE)

#> [1] TRUE
#> 

testCharacter(letters, min.chars = 2)

#> [1] FALSE
#> 

testCharacter("example", pattern = "xa")

#> [1] TRUE
#>