Check if an object is an element of a given set

Check if an object is an element of a given set

checkChoice(x, choices)

check_choice(x, choices)

assertChoice(x, choices, .var.name = vname(x), add = NULL)

assert_choice(x, choices, .var.name = vname(x), add = NULL)

testChoice(x, choices)

test_choice(x, choices)

expect_choice(x, choices, info = NULL, label = vname(x))

Arguments

x

[any] Object to check.

choices

[atomic] Set of possible values.

.var.name

[character(1)] Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add

[AssertCollection] Collection to store assertion messages. See AssertCollection.

info

[character(1)] Extra information to be included in the message for the testthat reporter. See expect_that.

label

[character(1)] Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertChoice/assert_choice return x invisibly, whereas checkChoice/check_choice and testChoice/test_choice return TRUE. If the check is not successful, assertChoice/assert_choice throws an error message, testChoice/test_choice returns FALSE, and checkChoice returns a string with the error message. The function expect_choice always returns an expectation.

Note

The object x must be of the same type as the set w.r.t. typeof. Integers and doubles are both treated as numeric.

See also

Other set: checkSetEqual, checkSubset

Examples

testChoice("x", letters)

#> [1] TRUE
#> 

# x is not converted before the comparison (except for numerics)
testChoice(factor("a"), "a")

#> [1] FALSE
#> 

testChoice(1, "1")

#> [1] FALSE
#> 

testChoice(1, as.integer(1))

#> [1] TRUE
#>