checkConstant {checkmate}	R Documentation

Check if all elements of a vector are equal / constant

Description

Note that zero-length input is treated as constant.

Usage

checkConstant(x, tol = .Machine$double.eps^0.5)

assertConstant(x, tol = .Machine$double.eps^0.5, .var.name)

testConstant(x, tol = .Machine$double.eps^0.5)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

tol
[double(1)]
Numerical tolerance used if x is of type double or complex. Default is sqrt(.Machine$double.eps).

Value

Depending on the function prefix: If the check is successfull, all functions return TRUE. If the check is not successfull, assertConstant throws an error message, testConstant returns FALSE and checkConstant returns a string with the error message.

See Also

Other constant: assertInconstant, checkInconstant, testInconstant

Examples

testConstant(c(0, 1-0.9-0.1))

## [1] TRUE

 testConstant(c(0, 1-0.9-0.1), tol = 0)

## [1] FALSE

[Package checkmate version 0.9 Index]