Check if an argument is a data frame

Check if an argument is a data frame

checkDataFrame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

check_data_frame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

assertDataFrame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

assert_data_frame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

testDataFrame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

test_data_frame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

expect_data_frame(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE,
  info = NULL, label = vname(x))

Arguments

x

[any] Object to check.

types

[character] Character vector of class names. Each list element must inherit from at least one of the provided types. The types “logical”, “integer”, “integerish”, “double”, “numeric”, “complex”, “character”, “factor”, “atomic”, “vector” “atomicvector”, “array”, “matrix”, “list”, “function”, “environment” and “null” are supported. For other types inherits is used as a fallback to check x's inheritance. Defaults to character(0) (no check).

any.missing

[logical(1)] Are missing values allowed? Default is TRUE.

all.missing

[logical(1)] Are matrices with only missing values allowed? Default is TRUE.

min.rows

[integer(1)] Minimum number of rows.

min.cols

[integer(1)] Minimum number of columns.

nrows

[integer(1)] Exact number of rows.

ncols

[integer(1)] Exact number of columns.

row.names

[character(1)] Check for row names. Default is “NULL” (no check). See checkNamed for possible values. Note that you can use checkSubset to check for a specific set of names.

col.names

[character(1)] Check for column names. Default is “NULL” (no check). See checkNamed for possible values. Note that you can use checkSubset to test for a specific set of names.

null.ok

[logical(1)] If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name

[character(1)] Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add

[AssertCollection] Collection to store assertion messages. See AssertCollection.

info

[character(1)] Extra information to be included in the message for the testthat reporter. See expect_that.

label

[character(1)] Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertDataFrame/assert_data_frame return x invisibly, whereas checkDataFrame/check_data_frame and testDataFrame/test_data_frame return TRUE. If the check is not successful, assertDataFrame/assert_data_frame throws an error message, testDataFrame/test_data_frame returns FALSE, and checkDataFrame returns a string with the error message. The function expect_data_frame always returns an expectation.

See also

Other basetypes: checkArray, checkAtomic, checkCharacter, checkComplex, checkDate, checkEnvironment, checkFactor, checkFunction, checkIntegerish, checkInteger, checkList, checkLogical, checkMatrix, checkNull, checkNumeric, checkVector

Other compound: checkArray, checkDataTable, checkMatrix, checkTibble

Examples

testDataFrame(iris)

#> [1] TRUE
#> 

testDataFrame(iris, types = c("numeric", "factor"), min.rows = 1, col.names = "named")

#> [1] TRUE
#>