checkDirectory {checkmate}	R Documentation

Check for existence and access rights of directories

Description

Check for existence and access rights of directories

Usage

checkDirectory(x, access = "")

assertDirectory(x, access = "", .var.name)

testDirectory(x, access = "", .var.name)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

access
[character]
Single string with characters ‘r’, ‘w’ and ‘x’ to force a check for read, write or execute access rights.

Value

Depending on the function prefix: If the check is successfull, all functions return TRUE. If the check is not successfull, assertDirectory throws an error message, testDirectory returns FALSE and checkDirectory returns a string with the error message.

See Also

Other filesystem: assertAccess, checkAccess, testAccess; assertFile, checkFile, testFile

Examples

# Is R's home directory readable?
 testDirectory(R.home(), "r")

## [1] TRUE

 # Is R's home directory writable?
 testDirectory(R.home(), "w")

## [1] FALSE

[Package checkmate version 0.9 Index]