Check if an argument is an environment

Check if an argument is an environment

checkEnvironment(x, contains = character(0L), null.ok = FALSE)

check_environment(x, contains = character(0L), null.ok = FALSE)

assertEnvironment(
  x,
  contains = character(0L),
  null.ok = FALSE,
  .var.name = vname(x),
  add = NULL
)

assert_environment(
  x,
  contains = character(0L),
  null.ok = FALSE,
  .var.name = vname(x),
  add = NULL
)

testEnvironment(x, contains = character(0L), null.ok = FALSE)

test_environment(x, contains = character(0L), null.ok = FALSE)

expect_environment(
  x,
  contains = character(0L),
  null.ok = FALSE,
  info = NULL,
  label = vname(x)
)

Arguments

x
[any]
Object to check.

contains
[character]
Vector of object names expected in the environment. Defaults to character(0).

null.ok
[logical(1)]
If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name
[character(1)]
Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add
[AssertCollection]
Collection to store assertion messages. See AssertCollection.

info
[character(1)]
Extra information to be included in the message for the testthat reporter. See expect_that.

label
[character(1)]
Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertEnvironment/assert_environment return x invisibly, whereas checkEnvironment/check_environment and testEnvironment/test_environment return TRUE. If the check is not successful, assertEnvironment/assert_environment throws an error message, testEnvironment/test_environment returns FALSE, and checkEnvironment returns a string with the error message. The function expect_environment always returns an expectation.

See also

Examples

ee = as.environment(list(a = 1))
testEnvironment(ee)

#> [1] TRUE

testEnvironment(ee, contains = "a")

#> [1] TRUE