checkIntegerish {checkmate}	R Documentation

Check if an object is an integerish vector

Description

Check if an object is an integerish vector

Usage

checkIntegerish(x, tol = sqrt(.Machine$double.eps), lower = -Inf,
  upper = Inf, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

assertIntegerish(x, tol = sqrt(.Machine$double.eps), lower = -Inf,
  upper = Inf, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  .var.name)

testIntegerish(x, tol = sqrt(.Machine$double.eps), lower = -Inf,
  upper = Inf, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

tol
[double(1)]
Numerical tolerance used to check whether a double or complex can be converted. Default is sqrt(.Machine$double.eps).

lower
[numeric(1)]
Lower value all elements of x must be greater than.

upper
[numeric(1)]
Upper value all elements of x must be lower than.

any.missing
[logical(1)]
Are vectors with missing values allowed? Default is TRUE.

all.missing
[logical(1)]
Are vectors with only missing values allowed? Default is TRUE.

len
[integer(1)]
Exact expected length of x.

min.len
[integer(1)]
Minimal length of x.

max.len
[integer(1)]
Maximal length of x.

unique
[logical(1)]
Must all values be unique? Default is FALSE.

names
[character(1)]
Check for names. See checkNamed for possible values. Default is “any” which performs no check at all.

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ NA_character_ and NaN.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertIntegerish throws an error message, testIntegerish returns FALSE and checkIntegerish returns a string with the error message.

See Also

Other basetypes: assertArray, checkArray, testArray; assertAtomicVector, checkAtomicVector, testAtomicVector; assertAtomic, checkAtomic, testAtomic; assertCharacter, checkCharacter, testCharacter; assertComplex, checkComplex, testComplex; assertDataFrame, checkDataFrame, testDataFrame; assertEnvironment, checkEnvironment, testEnvironment; assertFactor, checkFactor, testFactor; assertFunction, checkFunction, testFunction; assertInteger, checkInteger, testInteger; assertList, checkList, testList; assertLogical, checkLogical, testLogical; assertMatrix, checkMatrix, testMatrix; assertNumeric, checkNumeric, testNumeric; assertVector, checkVector, testVector

Examples

testIntegerish(1L)

## [1] TRUE

 testIntegerish(1.)

## [1] TRUE

 testIntegerish(1:2, lower = 1L, upper = 2L, any.missing = FALSE)

## [1] TRUE

[Package checkmate version 1.2 Index]