checkNull {checkmate}	R Documentation

Check if an argument is NULL

Description

Check if an argument is NULL

Usage

checkNull(x)

assertNull(x, .var.name)

testNull(x)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertNull throws an error message, testNull returns FALSE and checkNull returns a string with the error message.

Examples

testNull(NULL)

## [1] TRUE

 testNull(1)

## [1] FALSE

[Package checkmate version 1.2 Index]