Check that an argument is a vector of type numeric

Check that an argument is a vector of type numeric

checkNumeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, null.ok = FALSE)

check_numeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, null.ok = FALSE)

assertNumeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

assert_numeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

testNumeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, null.ok = FALSE)

test_numeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, null.ok = FALSE)

expect_numeric(x, lower = -Inf, upper = Inf, finite = FALSE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, null.ok = FALSE,
  info = NULL, label = vname(x))

Arguments

x

[any] Object to check.

lower

[numeric(1)] Lower value all elements of x must be greater than.

upper

[numeric(1)] Upper value all elements of x must be lower than.

finite

[logical(1)] Check for only finite values? Default is FALSE.

any.missing

[logical(1)] Are vectors with missing values allowed? Default is TRUE.

all.missing

[logical(1)] Are vectors with only missing values allowed? Default is TRUE.

len

[integer(1)] Exact expected length of x.

min.len

[integer(1)] Minimal length of x.

max.len

[integer(1)] Maximal length of x.

unique

[logical(1)] Must all values be unique? Default is FALSE.

names

[character(1)] Check for names. See checkNamed for possible values. Default is “any” which performs no check at all. Note that you can use checkSubset to check for a specific set of names.

null.ok

[logical(1)] If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name

[character(1)] Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add

[AssertCollection] Collection to store assertion messages. See AssertCollection.

info

[character(1)] Extra information to be included in the message for the testthat reporter. See expect_that.

label

[character(1)] Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertNumeric/assert_numeric return x invisibly, whereas checkNumeric/check_numeric and testNumeric/test_numeric return TRUE. If the check is not successful, assertNumeric/assert_numeric throws an error message, testNumeric/test_numeric returns FALSE, and checkNumeric returns a string with the error message. The function expect_numeric always returns an expectation.

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ NA_character_ and NaN.

See also

Other basetypes: checkArray, checkAtomicVector, checkAtomic, checkCharacter, checkComplex, checkDataFrame, checkDate, checkEnvironment, checkFactor, checkFunction, checkIntegerish, checkInteger, checkList, checkLogical, checkMatrix, checkNull, checkVector

Examples

testNumeric(1)

#> [1] TRUE

testNumeric(1, min.len = 1, lower = 0)

#> [1] TRUE