checkScalar {checkmate}	R Documentation

Check if an argument is a single atomic value

Description

Check if an argument is a single atomic value

Usage

checkScalar(x, na.ok = FALSE)

assertScalar(x, na.ok = FALSE, .var.name)

testScalar(x, na.ok = FALSE)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

na.ok
[logical(1)]
Are missing values allowed? Default is FALSE.

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ NA_character_ and NaN.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertScalar throws an error message, testScalar returns FALSE and checkScalar returns a string with the error message.

See Also

Other scalars: assertCount, checkCount, testCount; assertFlag, checkFlag, testFlag; assertInt, checkInt, testInt; assertNumber, checkNumber, testNumber; assertScalarNA, checkScalarNA, testScalarNA; assertString, checkString, testString

Examples

testScalar(1)

## [1] TRUE

[Package checkmate version 0.9 Index]