checkSetEqual {checkmate}	R Documentation

Check if object is a subset of a given set

Description

Check if object is a subset of a given set

Usage

checkSetEqual(x, y)

assertSetEqual(x, y, .var.name)

testSetEqual(x, y)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

y
[atomic]
Set to compare with.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertSubset throws an error message, testSubset returns FALSE and checkSubset returns a string with the error message.

See Also

Other set: assertChoice, checkChoice, testChoice; assertSubset, checkSubset, testSubset

Examples

testSetEqual(c("a", "b"), c("a", "b"))

## [1] TRUE

 testSetEqual(1:3, 1:4)

## [1] FALSE

[Package checkmate version 1.2 Index]