asInteger {checkmate}	R Documentation

Convert an argument to an integer

Description

Convertible are (a) atomic vectors with all elements NA and (b) double vectors with all elements being within tol range of an integer.

Usage

asInteger(x, ..., tol = .Machine$double.eps^0.5, .var.name)

Arguments

x
[ANY]
Object to convert.

...
[ANY]
Additional arguments passed to checkInteger.

tol
[double(1)]
Numerical tolerance used to check if a double or complex can be converted. Default is sqrt(.Machine$double.eps).

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

Value

Converted x.

[Package checkmate version 0.9 Index]