asInteger {checkmate}	R Documentation

Convert an argument to an integer

Description

asInteger is intended to be used for vectors while asInt is a specialization for scalar integers and asCount for scalar non-negative integers. Convertible are (a) atomic vectors with all elements NA and (b) double vectors with all elements being within tol range of an integer.

Usage

asInteger(x, ..., tol = .Machine$double.eps^0.5, .var.name)

asCount(x, na.ok = FALSE, positive = FALSE, .var.name)

asInt(x, na.ok = FALSE, lower = -Inf, upper = Inf, .var.name)

Arguments

x
[ANY]
Object to convert.

...
[ANY]
Additional arguments passed to checkInteger.

tol
[double(1)]
Numerical tolerance used to check whether a double or complex can be converted. Default is sqrt(.Machine$double.eps).

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

positive
[logical(1)]
Must x be positive (>= 1)? Default is FALSE.

na.ok
[logical(1)]
Are missing values allowed? Default is FALSE.

lower
[numeric(1)]
Lower value all elements of x must be greater than.

upper
[numeric(1)]
Upper value all elements of x must be lower than.

Value

Converted x.

Examples

asInteger(c(1, 2, 3))

## [1] 1 2 3

 asCount(1)

## [1] 1

 asInt(1)

## [1] 1

[Package checkmate version 0.9 Index]