assert {checkmate}	R Documentation

Combine multiple checks to an assertion

Description

Combine multiple checks to an assertion

Usage

assert(..., .var.name)

Arguments

...
[ANY]
List of calls to check functions.

.var.name
[character(1)]
Name for object to check. Defaults to a heuristic to determine the name of the first call.

Value

Throws an error if all checks failed and invisibly returns TRUE otherwise.

Examples

x = 1:10
 assert(checkNull(x), checkInteger(x, any.missing = FALSE))

## [1] TRUE

[Package checkmate version 0.9 Index]