checkAccess {checkmate}	R Documentation

Check file system access rights

Description

Check file system access rights

Usage

checkAccess(x, access = "")

assertAccess(x, access = "", .var.name)

testAccess(x, access = "")

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

access
[character]
Single string with characters ‘r’, ‘w’ and ‘x’ to force a check for read, write or execute access rights.

Value

Depending on the function prefix: If the check is successfull, all functions return TRUE. If the check is not successfull, assertAccess throws an error message, testAccess returns FALSE and checkAccess returns a string with the error message.

See Also

Other filesystem: assertDirectory, checkDirectory, testDirectory; assertFile, checkFile, testFile

Examples

# Is R's home directory readable?
 testAccess(R.home(), "r")

## [1] TRUE

 # Is R's home directory writeable?
 testAccess(R.home(), "w")

## [1] FALSE

[Package checkmate version 0.9 Index]