Check if an argument is an array

Check if an argument is an array

checkArray(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE)

check_array(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE)

assertArray(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

assert_array(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

testArray(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE)

test_array(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE)

expect_array(x, mode = NULL, any.missing = TRUE, d = NULL,
  min.d = NULL, max.d = NULL, null.ok = FALSE, info = NULL,
  label = vname(x))

Arguments

x
[any]
Object to check.

mode
[character(1)]
Storage mode of the array. Arrays can hold vectors, i.e. “logical”, “integer”, “integerish”, “double”, “numeric”, “complex”, “character” and “list”. You can also specify “atomic” here to explicitly prohibit lists. Default is NULL (no check).

any.missing
[logical(1)]
Are missing values allowed? Default is TRUE.

d
[integer(1)]
Exact number of dimensions of array x. Default is NULL (no check).

min.d
[integer(1)]
Minimum number of dimensions of array x. Default is NULL (no check).

max.d
[integer(1)]
Maximum number of dimensions of array x. Default is NULL (no check).

null.ok
[logical(1)]
If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name
[character(1)]
Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add
[AssertCollection]
Collection to store assertion messages. See AssertCollection.

info
[character(1)]
Extra information to be included in the message for the testthat reporter. See expect_that.

label
[character(1)]
Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertArray/assert_array return x invisibly, whereas checkArray/check_array and testArray/test_array return TRUE. If the check is not successful, assertArray/assert_array throws an error message, testArray/test_array returns FALSE, and checkArray returns a string with the error message. The function expect_array always returns an expectation.

See also

Examples

checkArray(array(1:27, dim = c(3, 3, 3)), d = 3)

#> [1] TRUE