checkCharacter {checkmate}	R Documentation

Check if an argument is a vector of type character

Description

Check if an argument is a vector of type character

Usage

checkCharacter(x, min.chars = NULL, pattern = NULL, fixed = FALSE,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL)

assertCharacter(x, min.chars = NULL, pattern = NULL, fixed = FALSE,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, .var.name)

testCharacter(x, min.chars = NULL, pattern = NULL, fixed = FALSE,
  ignore.case = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

pattern
[character(1L)]
Regular expression as used in grepl. All elements of x must comply to this pattern. Defaults to NULL.

ignore.case
[logical(1)]
See grepl. Default is FALSE.

fixed
[logical(1)]
See grepl. Default is FALSE.

min.chars
[integer(1)]
Minimum number of characters in each element of x.

any.missing
[logical(1)]
Are vectors with missing values allowed? Default is TRUE.

all.missing
[logical(1)]
Are vectors with only missing values allowed? Default is TRUE.

len
[integer(1)]
Exact expected length of x.

min.len
[integer(1)]
Minimal length of x.

max.len
[integer(1)]
Maximal length of x.

unique
[logical(1)]
Must all values be unique? Default is FALSE.

names
[character(1)]
Check for names. See checkNamed for possible values. Default is “any” which performs no check at all.

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ NA_character_ and NaN.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertCharacter throws an error message, testCharacter returns FALSE and checkCharacter returns a string with the error message.

See Also

Other basetypes: assertArray, checkArray, testArray; assertAtomicVector, checkAtomicVector, testAtomicVector; assertAtomic, checkAtomic, testAtomic; assertComplex, checkComplex, testComplex; assertDataFrame, checkDataFrame, testDataFrame; assertEnvironment, checkEnvironment, testEnvironment; assertFactor, checkFactor, testFactor; assertFunction, checkFunction, testFunction; assertIntegerish, checkIntegerish, testIntegerish; assertInteger, checkInteger, testInteger; assertList, checkList, testList; assertLogical, checkLogical, testLogical; assertMatrix, checkMatrix, testMatrix; assertNumeric, checkNumeric, testNumeric; assertVector, checkVector, testVector

Examples

testCharacter(letters, min.len = 1, any.missing = FALSE)

## [1] TRUE

 testCharacter(letters, min.chars = 2)

## [1] FALSE

 testCharacter("example", pattern = "xa")

## [1] TRUE

[Package checkmate version 1.2 Index]