checkCount {checkmate}	R Documentation

Check if an argument is a count

Description

A count is a non-negative integerish value.

Usage

checkCount(x, na.ok = FALSE, positive = FALSE,
  tol = sqrt(.Machine$double.eps))

assertCount(x, na.ok = FALSE, positive = FALSE,
  tol = sqrt(.Machine$double.eps), .var.name)

testCount(x, na.ok = FALSE, positive = FALSE,
  tol = sqrt(.Machine$double.eps))

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

na.ok
[logical(1)]
Are missing values allowed? Default is FALSE.

positive
[logical(1)]
Must x be positive (>= 1)? Default is FALSE, allowing 0.

tol
[double(1)]
Numerical tolerance used to check whether a double or complex can be converted. Default is sqrt(.Machine$double.eps).

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ NA_character_ and NaN.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertCount throws an error message, testCount returns FALSE and checkCount returns a string with the error message.

See Also

Other scalars: assertFlag, checkFlag, testFlag; assertInt, checkInt, testInt; assertNumber, checkNumber, testNumber; assertScalarNA, checkScalarNA, testScalarNA; assertScalar, checkScalar, testScalar; assertString, checkString, testString

Examples

testCount(1)

## [1] TRUE

 testCount(-1)

## [1] FALSE

[Package checkmate version 1.0 Index]