checkDataFrame {checkmate}	R Documentation

Check if an argument is a data frame

Description

Check if an argument is a data frame

Usage

checkDataFrame(x, types = character(0L), any.missing = TRUE,
  min.rows = NULL, min.cols = NULL, nrows = NULL, ncols = NULL,
  row.names = NULL, col.names = NULL)

assertDataFrame(x, types = character(0L), any.missing = TRUE,
  min.rows = NULL, min.cols = NULL, nrows = NULL, ncols = NULL,
  row.names = NULL, col.names = NULL, .var.name)

testDataFrame(x, types = character(0L), any.missing = TRUE,
  min.rows = NULL, min.cols = NULL, nrows = NULL, ncols = NULL,
  row.names = NULL, col.names = NULL)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

any.missing
[logical(1)]
Are missing values allowed? Default is TRUE.

min.rows
[integer(1)]
Minimum number of rows.

min.cols
[integer(1)]
Minimum number of columns.

nrows
[integer(1)]
Exact number of rows.

ncols
[integer(1)]
Exact number of columns.

row.names
[character(1)]
Check for row names. Default is “NULL” (no check). See checkNamed for possible values.

col.names
[character(1)]
Check for column names. Default is “NULL” (no check). See checkNamed for possible values.

types
[character]
Character vector of class names. Each list element must inherit from at least one of the provided types. The types “logical”, “integer”, “integerish”, “double”, “numeric”, “complex”, “character”, “factor”, “atomic”, “vector” “atomicvector”, “array”, “matrix”, “list” and “null” are supported. For other types inherits is used as a fallback to check x's inheritance. Defaults to character(0) (no check).

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertDataFrame throws an error message, testDataFrame returns FALSE and checkDataFrame returns a string with the error message.

See Also

Other basetypes: assertArray, checkArray, testArray; assertAtomicVector, checkAtomicVector, testAtomicVector; assertAtomic, checkAtomic, testAtomic; assertCharacter, checkCharacter, testCharacter; assertComplex, checkComplex, testComplex; assertEnvironment, checkEnvironment, testEnvironment; assertFactor, checkFactor, testFactor; assertFunction, checkFunction, testFunction; assertIntegerish, checkIntegerish, testIntegerish; assertInteger, checkInteger, testInteger; assertList, checkList, testList; assertLogical, checkLogical, testLogical; assertMatrix, checkMatrix, testMatrix; assertNumeric, checkNumeric, testNumeric; assertVector, checkVector, testVector

Examples

testDataFrame(iris, "data.frame")

## [1] FALSE

 testDataFrame(iris, "data.frame", min.rows = 1, col.names = "named")

## [1] FALSE

[Package checkmate version 1.0 Index]