Check that an argument is a Date

Checks that an object is of class Date.

checkDate(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE)

check_date(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE)

assertDate(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE, .var.name = vname(x),
  add = NULL)

assert_date(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE, .var.name = vname(x),
  add = NULL)

testDate(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE)

test_date(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE)

expect_date(x, lower = NULL, upper = NULL, any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, null.ok = FALSE, info = NULL, label = vname(x))

Arguments

x
[any]
Object to check.

lower
[Date]
All non-missing dates in x must be >= this date. Comparison is done via Ops.Date.

upper
[Date]
All non-missing dates in x must be before <= this date. Comparison is done via Ops.Date.

any.missing
[logical(1)]
Are vectors with missing values allowed? Default is TRUE.

all.missing
[logical(1)]
Are vectors with only missing values allowed? Default is TRUE.

len
[integer(1)]
Exact expected length of x.

min.len
[integer(1)]
Minimal length of x.

max.len
[integer(1)]
Maximal length of x.

unique
[logical(1)]
Must all values be unique? Default is FALSE.

null.ok
[logical(1)]
If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name
[character(1)]
Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add
[AssertCollection]
Collection to store assertion messages. See AssertCollection.

info
[character(1)]
Extra information to be included in the message for the testthat reporter. See expect_that.

label
[character(1)]
Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertAtomic/assert_atomic return x invisibly, whereas checkAtomic/check_atomic and testAtomic/test_atomic return TRUE. If the check is not successful, assertAtomic/assert_atomic throws an error message, testAtomic/test_atomic returns FALSE, and checkAtomic returns a string with the error message. The function expect_atomic always returns an expectation.

See also