checkFactor {checkmate}	R Documentation

Check if an argument is a factor

Description

Check if an argument is a factor

Usage

checkFactor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL)

assertFactor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL, .var.name)

testFactor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, unique = FALSE, names = NULL)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

levels
[character]
Vector of allowed factor levels.

ordered
[logical(1)]
Check for an ordered factor? If FALSE or TRUE, checks explicitly for an unordered or ordered factor, respectively. Default is NA which disables this check.

empty.levels.ok
[logical(1)]
Are empty levels allowed? Default is TRUE.

any.missing
[logical(1)]
Are vectors with missing values allowed? Default is TRUE.

all.missing
[logical(1)]
Are vectors with only missing values allowed? Default is TRUE.

len
[integer(1)]
Exact expected length of x.

min.len
[integer(1)]
Minimal length of x.

max.len
[integer(1)]
Maximal length of x.

unique
[logical(1)]
Must all values be unique? Default is FALSE.

names
[character(1)]
Check for names. See checkNamed for possible values. Default is “any” which performs no check at all.

Value

Depending on the function prefix: If the check is successfull, all functions return TRUE. If the check is not successfull, assertFactor throws an error message, testFactor returns FALSE and checkFactor returns a string with the error message.

See Also

Other basetypes: assertArray, checkArray, testArray; assertCharacter, checkCharacter, testCharacter; assertComplex, checkComplex, testComplex; assertDataFrame, checkDataFrame, testDataFrame; assertEnvironment, checkEnvironment, testEnvironment; assertFunction, checkFunction, testFunction; assertIntegerish, checkIntegerish, testIntegerish; assertInteger, checkInteger, testInteger; assertList, checkList, testList; assertLogical, checkLogical, testLogical; assertMatrix, checkMatrix, testMatrix; assertNumeric, checkNumeric, testNumeric; assertVector, checkVector, testVector

Examples

x = factor("a", levels = c("a", "b"))
 testFactor(x)

## [1] TRUE

 testFactor(x, empty.levels.ok = FALSE)

## [1] FALSE

[Package checkmate version 0.9 Index]