Check if an argument is a factor

Check if an argument is a factor

checkFactor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, n.levels = NULL, min.levels = NULL, max.levels = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

check_factor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, n.levels = NULL, min.levels = NULL, max.levels = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

assertFactor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, n.levels = NULL, min.levels = NULL, max.levels = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE, .var.name = vname(x),
  add = NULL)

assert_factor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, n.levels = NULL, min.levels = NULL, max.levels = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE, .var.name = vname(x),
  add = NULL)

testFactor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, n.levels = NULL, min.levels = NULL, max.levels = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

test_factor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, n.levels = NULL, min.levels = NULL, max.levels = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

expect_factor(x, levels = NULL, ordered = NA, empty.levels.ok = TRUE,
  any.missing = TRUE, all.missing = TRUE, len = NULL, min.len = NULL,
  max.len = NULL, n.levels = NULL, min.levels = NULL, max.levels = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE, info = NULL,
  label = vname(x))

Arguments

x

[any] Object to check.

levels

[character] Vector of allowed factor levels.

ordered

[logical(1)] Check for an ordered factor? If FALSE or TRUE, checks explicitly for an unordered or ordered factor, respectively. Default is NA which does not perform any additional check.

empty.levels.ok

[logical(1)] Are empty levels allowed? Default is TRUE.

any.missing

[logical(1)] Are vectors with missing values allowed? Default is TRUE.

all.missing

[logical(1)] Are vectors with only missing values allowed? Default is TRUE.

len

[integer(1)] Exact expected length of x.

min.len

[integer(1)] Minimal length of x.

max.len

[integer(1)] Maximal length of x.

n.levels

[integer(1)] Exact number of factor levels. Default is NULL (no check).

min.levels

[integer(1)] Minimum number of factor levels. Default is NULL (no check).

max.levels

[integer(1)] Maximum number of factor levels. Default is NULL (no check).

unique

[logical(1)] Must all values be unique? Default is FALSE.

names

[character(1)] Check for names. See checkNamed for possible values. Default is “any” which performs no check at all. Note that you can use checkSubset to check for a specific set of names.

null.ok

[logical(1)] If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name

[character(1)] Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add

[AssertCollection] Collection to store assertion messages. See AssertCollection.

info

[character(1)] Extra information to be included in the message for the testthat reporter. See expect_that.

label

[character(1)] Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertFactor/assert_factor return x invisibly, whereas checkFactor/check_factor and testFactor/test_factor return TRUE. If the check is not successful, assertFactor/assert_factor throws an error message, testFactor/test_factor returns FALSE, and checkFactor returns a string with the error message. The function expect_factor always returns an expectation.

See also

Other basetypes: checkArray, checkAtomic, checkCharacter, checkComplex, checkDataFrame, checkDate, checkEnvironment, checkFunction, checkIntegerish, checkInteger, checkList, checkLogical, checkMatrix, checkNull, checkNumeric, checkVector

Examples

x = factor("a", levels = c("a", "b"))
testFactor(x)

#> [1] TRUE
#> 

testFactor(x, empty.levels.ok = FALSE)

#> [1] FALSE
#>