checkInt {checkmate}	R Documentation

Check if an argument is a single integerish value

Description

Check if an argument is a single integerish value

Usage

checkInt(x, na.ok = FALSE, lower = -Inf, upper = Inf,
  tol = .Machine$double.eps^0.5)

assertInt(x, na.ok = FALSE, lower = -Inf, upper = Inf,
  tol = .Machine$double.eps^0.5, .var.name)

testInt(x, na.ok = FALSE, lower = -Inf, upper = Inf,
  tol = .Machine$double.eps^0.5)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

na.ok
[logical(1)]
Are missing values allowed? Default is FALSE.

lower
[numeric(1)]
Lower value all elements of x must be greater than.

upper
[numeric(1)]
Upper value all elements of x must be lower than.

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ NA_character_ and NaN.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertInt throws an error message, testInt returns FALSE and checkInt returns a string with the error message.

See Also

Other scalars: assertCount, checkCount, testCount; assertFlag, checkFlag, testFlag; assertNumber, checkNumber, testNumber; assertScalarNA, checkScalarNA, testScalarNA; assertScalar, checkScalar, testScalar; assertString, checkString, testString

Examples

testInt(1)

## [1] TRUE

 testInt(-1, lower = 0)

## [1] FALSE

[Package checkmate version 0.9 Index]