Check if an argument is a list

Check if an argument is a list

checkList(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

check_list(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

assertList(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

assert_list(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

testList(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

test_list(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE)

expect_list(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
  unique = FALSE, names = NULL, null.ok = FALSE, info = NULL,
  label = vname(x))

Arguments

x
[any]
Object to check.

types
[character]
Character vector of class names. Each list element must inherit from at least one of the provided types. The types “logical”, “integer”, “integerish”, “double”, “numeric”, “complex”, “character”, “factor”, “atomic”, “vector” “atomicvector”, “array”, “matrix”, “list”, “function”, “environment” and “null” are supported. For other types inherits is used as a fallback to check x's inheritance. Defaults to character(0) (no check).

any.missing
[logical(1)]
Are vectors with missing values allowed? Default is TRUE.

all.missing
[logical(1)]
Are vectors with only missing values allowed? Default is TRUE.

len
[integer(1)]
Exact expected length of x.

min.len
[integer(1)]
Minimal length of x.

max.len
[integer(1)]
Maximal length of x.

unique
[logical(1)]
Must all values be unique? Default is FALSE.

names
[character(1)]
Check for names. See checkNamed for possible values. Default is “any” which performs no check at all. Note that you can use checkSubset to check for a specific set of names.

null.ok
[logical(1)]
If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name
[character(1)]
Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add
[AssertCollection]
Collection to store assertion messages. See AssertCollection.

info
[character(1)]
Extra information to be included in the message for the testthat reporter. See expect_that.

label
[character(1)]
Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

...
[any]
Additional parameters used in a call of checkVector.

Value

Depending on the function prefix: If the check is successful, the functions assertList/assert_list return x invisibly, whereas checkList/check_list and testList/test_list return TRUE. If the check is not successful, assertList/assert_list throws an error message, testList/test_list returns FALSE, and checkList returns a string with the error message. The function expect_list always returns an expectation.

Note

Contrary to R's is.list, objects of type pairlist are not recognized as list.

Missingness is defined here as elements of the list being NULL, analogously to anyMissing.

The test for uniqueness does differentiate between the different NA types which are built-in in R. This is required to be consistent with unique while checking scalar missing values. Also see the example.

See also

Examples

testList(list())

#> [1] TRUE

testList(as.list(iris), types = c("numeric", "factor"))

#> [1] TRUE

# Missingness
testList(list(1, NA), any.missing = FALSE)

#> [1] TRUE

testList(list(1, NULL), any.missing = FALSE)

#> [1] FALSE

# Uniqueness differentiates between different NA types:
testList(list(NA, NA), unique = TRUE)

#> [1] FALSE

testList(list(NA, NA_real_), unique = TRUE)

#> [1] TRUE