checkMatrix {checkmate}	R Documentation

Check if an argument is a matrix

Description

Check if an argument is a matrix

Usage

checkMatrix(x, mode = "any", any.missing = TRUE, min.rows = NULL,
  min.cols = NULL, nrows = NULL, ncols = NULL, row.names = NULL,
  col.names = NULL)

assertMatrix(x, mode = "any", any.missing = TRUE, min.rows = NULL,
  min.cols = NULL, nrows = NULL, ncols = NULL, row.names = NULL,
  col.names = NULL, .var.name)

testMatrix(x, mode = "any", any.missing = TRUE, min.rows = NULL,
  min.cols = NULL, nrows = NULL, ncols = NULL, row.names = NULL,
  col.names = NULL)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

mode
[character(1)]
Storage mode of the matrix. Matricies can hold “logical”, “integer”, “double”, “complex” and “character”. “numeric” is also supported by this function. Default is “any”.

any.missing
[logical(1)]
Are missing values allowed? Default is TRUE.

min.rows
[integer(1)]
Minimum number of rows.

min.cols
[integer(1)]
Minimum number of columns.

nrows
[integer(1)]
Exact number of rows.

ncols
[integer(1)]
Exact number of columns.

row.names
[character(1)]
Check for row names. Default is “any” (no check). See checkNamed for possible values.

col.names
[character(1)]
Check for column names. Default is “any” (no check). See checkNamed for possible values.

Value

Depending on the function prefix: If the check is successfull, all functions return TRUE. If the check is not successfull, assertMatrix throws an error message, testMatrix returns FALSE and checkMatrix returns a string with the error message.

See Also

Other basetypes: assertArray, checkArray, testArray; assertCharacter, checkCharacter, testCharacter; assertComplex, checkComplex, testComplex; assertDataFrame, checkDataFrame, testDataFrame; assertEnvironment, checkEnvironment, testEnvironment; assertFactor, checkFactor, testFactor; assertFunction, checkFunction, testFunction; assertIntegerish, checkIntegerish, testIntegerish; assertInteger, checkInteger, testInteger; assertList, checkList, testList; assertLogical, checkLogical, testLogical; assertNumeric, checkNumeric, testNumeric; assertVector, checkVector, testVector

Examples

x = matrix(1:9, 3)
 colnames(x) = letters[1:3]
 testMatrix(x, nrows = 3, min.cols = 1, col.names = "named")

## [1] TRUE

[Package checkmate version 0.9 Index]