checkNumeric {checkmate}	R Documentation

Check that an argument is a vector of type numeric

Description

Check that an argument is a vector of type numeric

Usage

checkNumeric(x, lower = -Inf, upper = Inf, ...)

assertNumeric(x, lower = -Inf, upper = Inf, ..., .var.name)

testNumeric(x, lower = -Inf, upper = Inf, ...)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

lower
[numeric(1)]
Lower value all elements of x must be greater than.

upper
[numeric(1)]
Upper value all elements of x must be lower than.

...
[ANY]
Additional parameters used in a call of checkVector.

Details

This function does not distinguish between NA, NA_integer_, NA_real_, NA_complex_ and NA_character_.

Value

Depending on the function prefix: If the check is successfull, all functions return TRUE. If the check is not successfull, assertNumeric throws an error message, testNumeric returns FALSE and checkNumeric returns a string with the error message.

See Also

Other basetypes: assertCharacter, checkCharacter, testCharacter; assertComplex, checkComplex, testComplex; assertDataFrame, checkDataFrame, testDataFrame; assertEnvironment, checkEnvironment, testEnvironment; assertFactor, checkFactor, testFactor; assertFunction, checkFunction, testFunction; assertIntegerish, checkIntegerish, testIntegerish; assertInteger, checkInteger, testInteger; assertList, checkList, testList; assertLogical, checkLogical, testLogical; assertMatrix, checkMatrix, testMatrix; assertVector, checkVector, testVector

Examples

testNumeric(1)

## [1] TRUE

 testNumeric(1, min.len = 1, lower = 0)

## [1] TRUE

[Package checkmate version 0.9 Index]