checkPathForOutput {checkmate}	R Documentation

Check file path for later output

Description

Check whether a file path can later be safely used to create a file and write to it.

This is checked:

Does dirname(x) exist?

Does no file under path x) exist?

Is dirname(x) writeable?

A string without slashes is interpreted as a file in the current working directory.

Usage

checkPathForOutput(x)

assertPathForOutput(x, .var.name)

testPathForOutput(x)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertPathForOutput throws an error message, testPathForOutput returns FALSE and checkPathForOutput returns a string with the error message.

See Also

Other filesystem: assertAccess, checkAccess, testAccess; assertDirectory, checkDirectory, testDirectory; assertFile, checkFile, testFile

Examples

# Can we create a file in the tempdir?
testPathForOutput(file.path(tempdir(), "process.log"))

## [1] TRUE

[Package checkmate version 1.2 Index]