checkSubset {checkmate}	R Documentation

Check if object is a subset of a given set

Description

Check if object is a subset of a given set

Usage

checkSubset(x, choices)

assertSubset(x, choices, .var.name)

testSubset(x, choices)

Arguments

x
[ANY]
Object to check.

.var.name
[character(1)]
Name for x. Defaults to a heuristic to determine the name using deparse and substitute.

choices
[atomic]
Set of possible values.

Value

Depending on the function prefix: If the check is successful, all functions return TRUE. If the check is not successful, assertSubset throws an error message, testSubset returns FALSE and checkSubset returns a string with the error message.

See Also

Other set: assertChoice, checkChoice, testChoice

Examples

testSubset(c("a", "z"), letters)

## [1] TRUE

[Package checkmate version 0.9 Index]