Check if an argument is a vector

Check if an argument is a vector

checkVector(x, strict = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

check_vector(x, strict = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

assertVector(x, strict = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE, .var.name = vname(x), add = NULL)

assert_vector(x, strict = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE, .var.name = vname(x), add = NULL)

testVector(x, strict = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

test_vector(x, strict = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE)

expect_vector(x, strict = FALSE, any.missing = TRUE, all.missing = TRUE,
  len = NULL, min.len = NULL, max.len = NULL, unique = FALSE,
  names = NULL, null.ok = FALSE, info = NULL, label = vname(x))

Arguments

x
[any] Object to check.

strict
[logical(1)] May the vector have additional attributes or perform a check for additional attributes like is.vector? Default is FALSE which allows e.g. factors or data.frames to be recognized as vectors.

any.missing
[logical(1)] Are vectors with missing values allowed? Default is TRUE.

all.missing
[logical(1)] Are vectors with only missing values allowed? Default is TRUE.

len
[integer(1)] Exact expected length of x.

min.len
[integer(1)] Minimal length of x.

max.len
[integer(1)] Maximal length of x.

unique
[logical(1)] Must all values be unique? Default is FALSE.

names
[character(1)] Check for names. See checkNamed for possible values. Default is “any” which performs no check at all. Note that you can use checkSubset to check for a specific set of names.

null.ok
[logical(1)] If set to TRUE, x may also be NULL. In this case only a type check of x is performed, all additional checks are disabled.

.var.name
[character(1)] Name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add
[AssertCollection] Collection to store assertion messages. See AssertCollection.

info
[character(1)] Extra information to be included in the message for the testthat reporter. See expect_that.

label
[character(1)] Name of the checked object to print in messages. Defaults to the heuristic implemented in vname.

Value

Depending on the function prefix: If the check is successful, the functions assertVector/assert_vector return x invisibly, whereas checkVector/check_vector and testVector/test_vector return TRUE. If the check is not successful, assertVector/assert_vector throws an error message, testVector/test_vector returns FALSE, and checkVector returns a string with the error message. The function expect_vector always returns an expectation.

See also

Other basetypes: checkArray, checkAtomicVector, checkAtomic, checkCharacter, checkComplex, checkDataFrame, checkDate, checkEnvironment, checkFactor, checkFunction, checkIntegerish, checkInteger, checkList, checkLogical, checkMatrix, checkNull, checkNumeric

Other atomicvector: checkAtomicVector, checkAtomic

Examples

testVector(letters, min.len = 1L, any.missing = FALSE)

#> [1] TRUE