qassertr {checkmate}	R Documentation

Recursive arguments checks on lists and data frames.

Description

These functions are the tuned counterparts of qcheck and qassert tailored for recursive checks of list elements or data frame columns.

Usage

qassertr(x, rules, .var.name)

qcheckr(x, rules)

Arguments

x
[list or data.frame]
List or data frame to check for compliance with at least one of rules. See details of qcheck for rule explanation.

rules
[character]
Set of rules. See qcheck

.var.name
[logical(1)]
Argument name to print in error message. If missing, the name of x will be retrieved via substitute.

Value

[logical(1)]: TRUE on success, FALSE (or a thrown exception) otherwise.

See Also

qcheck, qassert

Examples

qcheckr(as.list(1:10), "i+")

## [1] TRUE

qcheck(iris, "n")

## [1] FALSE

[Package checkmate version 0.9 Index]